requirements.txt
pandas
numpy
google-genai
python-dotenv
python-docx
openpyxl
spyder-kernels==3.1.*
